Increased minimum Ansible version to 2.2.0.0
Reason was that PR #21 uses functionality from the `six` package,
and it seemed reasonable to use the version of `six` that is bundled
with the Ansible Python module support (`ansible.module_utils` package).
That drives the need to increase the Ansible minimum version from 2.0.0.1
to 2.2.0.0.

Analysis of the impact of this upgrade for some Linux distros:
* For RHEL/CentOS 7, Ansible 2.3 can be installed from the EPEL repo
  (see https://www.liquidweb.com/kb/how-to-install-ansible-on-
   centos-7-via-yum/).
* For Ubuntu, Ansible 2.3 can be installed from an Ansible PPA (see
  https://www.digitalocean.com/community/tutorials/how-to-install-and-
  configure-ansible-on-ubuntu-14-04).

Signed-off-by: Andreas Maier <maiera@de.ibm.com>