Change structure to ansible collection format.
* Convert structure to ansible collection format. Ansible collections are
a new format for distributing custom plugins(modules) for ansible. They
can be published on Ansible Galaxy and freely installed from there by
the end users. This change implements the required structural changes to
comply with ansible collection requirements.

* Due to required ansible collection folder structure, the existing method
of pip installation had to be changed to install the modules into the
modules subfolder of the zhmc_ansible_modules folder. This is a breaking
change for the existing installations.

* Ansible Galaxy requires ansible sanity tests to pass before a collection
can be published. This change also addresses the problems identified by the
sanity tests and adds automatic sanity testing to Travis CI.

* Convert the README.rst from reStructuredText to markdown as required by
Ansible Galaxy.

* Pin ansible version used in doc generation to 2.9.12 due to the API change
in ansible 2.10 (devel branch was used before).

* Cleanup: remove attic, get_facts.py and provision.yml which are not used.

Signed-off-by: Adilet Sabyrbaev <adilet.sabyrbaev2@ibm.com>